aggregation logic extracted from shplonk.rs
